+++ ESFramework/Foundation/NSData+ESExtension.h
@@ -19,6 +19,11 @@ NS_ASSUME_NONNULL_BEGIN
 + (nullable NSData *)dataWithHexString:(NSString *)hexString;
 
 /**
+ * Generates a random data using `SecRandomCopyBytes`.
+ */
++ (nullable NSData *)randomDataWithLength:(NSUInteger)length;
+
+/**
  * Converts the data to a UTF-8 encoded string.
  */
 - (NSString *)UTF8String;